Bản update có tính sát thương cực cao

+++ app/src/main/java/hecosodulieudaphuongtien/demonhom19/webservice/MyService.java
@@ -1,19 +1,8 @@
 package hecosodulieudaphuongtien.demonhom19.webservice;
 
-import com.google.gson.JsonObject;
-
-import java.util.Map;
-
-import okhttp3.RequestBody;
 import okhttp3.ResponseBody;
 import retrofit2.Call;
-import retrofit2.http.Body;
 import retrofit2.http.GET;
-import retrofit2.http.Headers;
-import retrofit2.http.Multipart;
-import retrofit2.http.POST;
-import retrofit2.http.Part;
-import retrofit2.http.PartMap;
 import retrofit2.http.Query;
 
 /**
@@ -37,6 +26,6 @@ public interface MyService {
     Call<ResponseBody> updateView(@Query("id") int id);
 
     @GET("AudioPhantan/api/update_Rate.php")
-    Call<ResponseBody> updateRate(@Query("id") int id, @Query("rate") int rate);
+    Call<ResponseBody> updateRate(@Query("id") int id, @Query("rate") float rate);
 
 }
